Provide slice publisher operator
