+++ Demo/Sources/Showcase/LinkView.swift
@@ -12,7 +12,7 @@ struct LinkView: View {
     let media: Media
 
     @StateObject private var player = Player(configuration: .externalPlaybackDisabled)
-    @StateObject private var isBusyTracker = PropertyTracker(keyPath: \.isBusy)
+    @StateObject private var isBusyTracker = PropertyTracker(at: \.isBusy)
     @State private var isDisplayed = true
 
     var body: some View {
